+++ arduino_tape_emulator/IMenuDrawer.h
@@ -5,8 +5,8 @@ class IMenuDrawer
 {
 public:
     virtual uint16_t maxItems() = 0;
-    virtual void drawItem(const string& text, size_t position, bool active) = 0;
-    virtual void quickDrawItem(const string& text, size_t position, bool active, bool fillAll) = 0;
+    virtual void drawItem(const string& text, uint16_t position, bool active) = 0;
+    virtual void quickDrawItem(const string& text, uint16_t position, bool active, bool fillAll) = 0;
 };
 
 #endif
